Nutrition Set

+++ app/src/main/java/org/d2e/smarthydroponic/data/Command.kt
@@ -4,6 +4,6 @@ data class Command (
     var flowPump: Int = 0,
     var maxTemp: Float = 0F,
     var minTemp: Float = 0F,
-    var nutriSet: Float = 0F,
+    var nutriSet: Int = 0,
     var phSet: Float = 0F
 )
